wrote methods to calculate distances to dispensaries
Currently, this only works on a dispensary's details page.  We intend to extend it to the dispensary cards as well.

In the process of coding this, I had to import the Google Maps geometry library.  I also added the ability to manually edit a dispensary's coordinates on the management page.